monoid-integrations/monoid-bigquery/requirements.txt
argcomplete==2.0.0
attrs==22.1.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
black==22.10.0
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
datamodel-code-generator==0.13.3
dnspython==2.2.1
email-validator==1.3.0
genson==1.2.2
google-auth==2.14.1
google-cloud-bigquery==3.4.0
idna==3.4
inflect==5.6.2
isodate==0.6.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
monoid-pydev==0.0.1a2
mypy-extensions==0.4.3
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
pathspec==0.10.1
platformdirs==2.5.2
prance==0.21.8.0
psycopg==3.1.4
psycopg-binary==3.1.4
pycodestyle==2.9.1
pydantic==1.10.2
pypika==0.48.9
pyrsistent==0.16.1
PySnooper==1.1.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semver==2.13.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
